requirements.txt
python-docx
comtypes